Get rid of unnecessary import and save logic
