+++ src/lib/components/ImportForm.svelte
@@ -1,9 +1,19 @@
 <script lang="ts">
-    import type { Conversation, Message } from "../types/content";
+    import type {
+        ChatGPTConversation,
+        ClaudeConversation,
+        Conversation,
+        Message,
+    } from "../types/content";
     import {
         conversations,
-        addNewConversationAndRefresh,
+        loadConversationsFromBackground,
+        addNewConversationsAndRefresh,
     } from "../state/conversations.svelte";
+    import {
+        convertChatGPTToDesiredFormat,
+        convertClaudeToDesiredFormat,
+    } from "../utils";
     let selectedFile: File | null = null;
     let jsonData: object;
     let selectedProvider: string = "chatgpt";
@@ -46,257 +56,11 @@
 
             console.log("Converted File content:", convertedConvs);
 
-            for (const convo of convertedConvs) {
-                chrome.runtime.sendMessage(
-                    {
-                        type: "SAVE_CONVERSATION",
-                        conversation: convo,
-                    },
-                    (response) => {
-                        if (response.success) {
-                            console.log("Conversation saved successfully");
-                        } else {
-                            console.error(
-                                "Failed to save conversation:",
-                                response.error,
-                            );
-                        }
-                    },
-                );
-            }
-
-            chrome.runtime.sendMessage(
-                {
-                    type: "GET_CONVERSATIONS",
-                },
-                async (response) => {
-                    console.log({ response });
-                    if (response.success) {
-                        await addNewConversationAndRefresh(
-                            response.conversations,
-                        );
-                        console.log("Conversations:", response.conversations);
-                    }
-                },
-            );
-            // Further processing of jsonData can happen here
+            await addNewConversationsAndRefresh(convertedConvs);
         } catch (error) {
             console.error("Error reading file:", error);
             fileWarning = "Failed to read the file.";
         }
-    }
-
-    // Define types for the Claude API response format
-    interface ClaudeContentBlock {
-        start_timestamp?: string;
-        stop_timestamp?: string;
-        type: string;
-        text?: string;
-        citations?: any[];
-    }
-
-    interface ClaudeChatMessage {
-        uuid: string;
-        text?: string;
-        content?: ClaudeContentBlock[];
-        sender: "human" | "assistant";
-        created_at: string;
-        updated_at: string;
-        attachments?: any[];
-        files?: any[];
-    }
-
-    interface ClaudeConversation {
-        uuid: string;
-        name: string;
-        created_at: string;
-        updated_at: string;
-        account?: {
-            uuid: string;
-        };
-        chat_messages: ClaudeChatMessage[];
-    }
-
-    /**
-     * Converts Claude API conversation data to the desired output format
-     * @param claudeData - Array of Claude conversation objects
-     * @returns Array of conversations in the desired format
-     */
-    function convertClaudeToDesiredFormat(
-        claudeData: ClaudeConversation[],
-    ): Conversation[] {
-        return claudeData.map((conversation) => {
-            // Extract basic conversation metadata
-            const result: Conversation = {
-                id: conversation.uuid,
-                title: conversation.name || "Untitled Conversation",
-                created_at: new Date(conversation.created_at).getTime(),
-                updated_at: new Date(conversation.updated_at).getTime(),
-                source: "Claude",
-                messages: [],
-            };
-
-            // Process messages
-            if (
-                conversation.chat_messages &&
-                Array.isArray(conversation.chat_messages)
-            ) {
-                result.messages = conversation.chat_messages.map((msg) => {
-                    // Determine message content
-                    let content = "";
-
-                    // Handle different content structures
-                    if (msg.text) {
-                        // Direct text field
-                        content = msg.text;
-                    } else if (msg.content && Array.isArray(msg.content)) {
-                        // Content array with text blocks
-                        content = msg.content
-                            .filter(
-                                (block) => block.type === "text" && block.text,
-                            )
-                            .map((block) => block.text)
-                            .join("\n");
-                    }
-
-                    const message: Message = {
-                        created_at: new Date(msg.created_at).getTime(),
-                        author: msg.sender === "human" ? "user" : "assistant",
-                        content: content,
-                    };
-
-                    return message;
-                });
-            }
-
-            return result;
-        });
-    }
-
-    interface ChatGPTMessage {
-        id: string;
-        author: {
-            role: "user" | "assistant" | "system";
-        };
-        create_time: number;
-        content: {
-            content_type: string;
-            parts?: string[];
-            text?: string;
-            thoughts?: Array<{
-                content: string;
-            }>;
-        };
-        metadata?: {
-            model_slug?: string;
-        };
-    }
-
-    interface ChatGPTConversation {
-        title: string;
-        create_time: number;
-        update_time: number;
-        mapping: {
-            [key: string]: {
-                message: ChatGPTMessage;
-                parent: string | null;
-                children: string[];
-            };
-        };
-        current_node: string;
-    }
-
-    function convertChatGPTToDesiredFormat(
-        chatGPTData: ChatGPTConversation[],
-    ): Conversation[] {
-        return chatGPTData.map((conversation) => {
-            // Extract basic conversation metadata
-            const result: Conversation = {
-                id:
-                    conversation.current_node ||
-                    Object.keys(conversation.mapping)[0],
-                title: conversation.title || "Untitled Conversation",
-                created_at: conversation.create_time * 1000, // Convert to milliseconds
-                updated_at: conversation.update_time * 1000,
-                source: "ChatGPT",
-                messages: [],
-            };
-
-            // Build message chain by following parent->child relationships
-            const messages: Message[] = [];
-            const visitedNodes = new Set<string>();
-
-            // Helper function to traverse the message tree
-            const traverseMessages = (nodeId: string) => {
-                if (visitedNodes.has(nodeId) || !conversation.mapping[nodeId])
-                    return;
-
-                visitedNodes.add(nodeId);
-                const node = conversation.mapping[nodeId];
-
-                if (node.message) {
-                    // Determine message content
-                    let content = "";
-                    const msgContent = node.message.content;
-
-                    if (msgContent.parts && Array.isArray(msgContent.parts)) {
-                        content = msgContent.parts.join("\n");
-                    } else if (msgContent.text) {
-                        content = msgContent.text;
-                    } else if (
-                        msgContent.thoughts &&
-                        Array.isArray(msgContent.thoughts)
-                    ) {
-                        content = msgContent.thoughts
-                            .map((t) => t.content)
-                            .join("\n");
-                    }
-
-                    if (content) {
-                        messages.push({
-                            created_at: node.message.create_time * 1000,
-                            author:
-                                node.message.author.role === "user"
-                                    ? "user"
-                                    : "assistant",
-                            content: content,
-                            // model: node.message.metadata?.model_slug,
-                        });
-                    }
-                }
-
-                // Process children
-                node.children.forEach((childId) => traverseMessages(childId));
-            };
-
-            // Find root nodes (nodes with no parent or parent not in mapping)
-            const rootNodes = Object.entries(conversation.mapping)
-                .filter(
-                    ([id, node]) =>
-                        !node.parent || !conversation.mapping[node.parent],
-                )
-                .map(([id]) => id);
-
-            // Traverse from all root nodes
-            rootNodes.forEach((rootId) => traverseMessages(rootId));
-
-            // Sort messages by timestamp (since tree traversal might not maintain order)
-            result.messages = messages.sort(
-                (a, b) => a.created_at - b.created_at,
-            );
-
-            return result;
-        });
-    }
-    /**
-     * Process a single Claude conversation
-     * @param conversation - Single Claude conversation object
-     * @returns Conversation in the desired format
-     */
-    function processSingleConversation(
-        conversation: ClaudeConversation,
-    ): Conversation {
-        return convertClaudeToDesiredFormat([conversation])[0];
     }
 
     /**
@@ -383,7 +147,7 @@
     </nav>
 </div>
 
-<style>
+<!-- <style>
     div {
         font-family: sans-serif;
         max-width: 600px;
@@ -466,4 +230,4 @@
         border-top: solid;
         border-bottom: solid;
     }
-</style>
+</style> -->
