--- src/options/import-main.ts
@@ -1,7 +1,0 @@
-import Import from './import.svelte';
-
-const app = new Import({
-  target: document.getElementById('app'),
-});
-
-export default app;
