--- src/options/import.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Import Data</title>
-    <!-- Svelte component styles will be injected here by Vite -->
-</head>
-<body>
-    <div id="app"></div>
-    <script type="module" src="./import-main.ts"></script>
-</body>
-</html>
