+++ src/options/options.svelte
@@ -1,11 +1,10 @@
 <script lang="ts">
-  // Future Svelte logic can go here
+    // Future Svelte logic can go here
 </script>
 
 <h1>Extension Options</h1>
 <p>This is the options page for the Zebra LLM Conversation Cache.</p>
-<a href="import.html">Import Data</a>
 
 <style>
-  /* Future Svelte-specific styles can go here */
+    /* Future Svelte-specific styles can go here */
 </style>
